Add integration test for Stratified W7 setup.
- setup method initializes the filepaths to the config
  file and atom data file, brings in the benchmark
  data and performs a run.

- methods to perform asserions for j_estimators and
  j_blue_estimators have been added.
