--- tardis/tests/data/w7_13d/tardis_w7.yml
@@ -1,39 +1,0 @@
-atom_data: kurucz_cd23_chianti_H_He.h5
-model:
-  abundances:
-    filename: tardis_w7_13d_abundances.dat
-    filetype: simple_ascii
-    type: file
-  structure:
-    filename: tardis_w7_13d_densities.dat
-    filetype: simple_ascii
-    type: file
-    v_inner_boundary: 12000.000 km/s
-    v_outer_boundary: 35000.000 km/s
-montecarlo:
-  black_body_sampling:
-    num: 1000000
-    start: 1 angstrom
-    stop: 1000000 angstrom
-  iterations: 20
-  last_no_of_packets: 2.0e+5
-  no_of_packets: 4.0e+4
-  no_of_virtual_packets: 10
-  nthreads: 2
-  seed: 23111963
-plasma:
-  initial_t_rad: 10000 K
-  initial_t_inner: 10000 K
-  disable_electron_scattering: false
-  excitation: dilute-lte
-  ionization: nebular
-  line_interaction_type: downbranch
-  radiative_rates_type: dilute-blackbody
-spectrum:
-  num: 10000
-  start: 500 angstrom
-  stop: 20000 angstrom
-supernova:
-  luminosity_requested: 9.44 log_lsun
-  time_explosion: 13. day
-tardis_config_version: v1.0
